moved ai assitance files out of gui
