+++ src/client/dcp_client/utils/ai_assistance/__init__.py
@@ -1,0 +1,12 @@
+"""AI assistance utilities (SAM controller, worker, preprocessing)."""
+
+from dcp_client.utils.ai_assistance.sam_controller import SAMController
+from dcp_client.utils.ai_assistance.sam_worker import SAMInferenceWorker
+from dcp_client.utils.ai_assistance.sam_preprocessing import preprocess_image, napari_rect_to_sam_box
+
+__all__ = [
+    "SAMController",
+    "SAMInferenceWorker",
+    "preprocess_image",
+    "napari_rect_to_sam_box",
+]
