+++ src/client/dcp_client/utils/ai_assistance/sam_worker.py
@@ -12,7 +12,7 @@ from typing import Optional, Tuple, List
 from PyQt5.QtCore import QThread, pyqtSignal, QMutex, QWaitCondition
 from queue import Queue, Empty
 
-from dcp_client.gui.sam_preprocessing import preprocess_image, napari_rect_to_sam_box
+from dcp_client.utils.ai_assistance.sam_preprocessing import preprocess_image, napari_rect_to_sam_box
 from dcp_client.utils.sam_model_manager import SAMModelManager
